+++ src-tauri/src/config.rs
@@ -67,6 +67,8 @@ pub struct Config {
   pub mkv: ConfigMkv,
   #[serde(rename = "batchMkvExtract", default)]
   pub batch_mkv_extract: ConfigBatchMkvExtract,
+  #[serde(rename = "bdMaster", default)]
+  pub bd_master: ConfigBDMaster,
   #[serde(default)]
   pub update: ConfigUpdate,
   #[serde(default)]
@@ -87,6 +89,7 @@ impl Default for Config {
       subtitle: Default::default(),
       mkv: Default::default(),
       batch_mkv_extract: Default::default(),
+      bd_master: Default::default(),
       update: Default::default(),
       window: Default::default(),
     }
@@ -215,6 +218,32 @@ impl ConfigBatchMkvExtract {
 }
 
 impl Default for ConfigBatchMkvExtract {
+  fn default() -> Self {
+    Self {
+      path: Self::default_path(),
+    }
+  }
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize)]
+pub struct ConfigBDMaster {
+  #[serde(default = "ConfigBDMaster::default_path")]
+  pub path: String,
+}
+
+impl ConfigBDMaster {
+  fn default_path() -> String {
+    if cfg!(target_os = "windows") {
+      r"C:\Program Files\BDMaster".to_owned()
+    } else if cfg!(target_os = "macos") {
+      "/Applications/BDMaster.app/Contents/MacOS".to_owned()
+    } else {
+      "/usr/bin".to_owned()
+    }
+  }
+}
+
+impl Default for ConfigBDMaster {
   fn default() -> Self {
     Self {
       path: Self::default_path(),