+++ docs/release_notes.md
@@ -4,6 +4,7 @@
 
 * Added "Open in MkvToolNix GUI" action to the card view for video files when MKVToolNix is configured.
 * Added multi-language support to the Windows installers.
+* Added Italian as a new supported language.
 
 ## 0.9.0